refactor(tailwind-integration): provide configuration through template
